# apidocs.yaml
openapi: 3.0.3
info:
  title: dockey api
  description: |-
    api docs for your favorite document collaborating app
  version: 1.0.0
servers:
  - url: http://74.242.214.25/v1/api
    description: Development server
...
paths:
...
  /documents:
    get:
      tags:
      - documents
      summary: Get all documents
      description: Retrieve a list of all documents
      operationId: getAllDocuments
      responses:
      '200':
        description: List of documents retrieved successfully
        content:
        application/json:
          schema:
          type: array
          items:
            $ref: '#/components/schemas/Document'
    post:
      tags:
      - documents
      summary: Create a new document
      description: Create a new document in the system
      operationId: createDocument
      requestBody:
      description: Document object to be created
      content:
        application/json:
        schema:
          $ref: '#/components/schemas/Document'
      required: true
      responses:
      '201':
        description: Document created successfully
        content:
        application/json:
          schema:
          $ref: '#/components/schemas/Document'
      '400':
        description: Invalid input
    /documents/{id}:
    get:
      tags:
      - documents
      summary: Get document by ID
      description: Retrieve a specific document by its ID with line comment counts
      operationId: getDocument
      parameters:
      - name: id
        in: path
        description: Document ID
        required: true
        schema:
        type: integer
        format: int64
      responses:
      '200':
        description: Document retrieved successfully
        content:
        application/json:
          schema:
          $ref: '#/components/schemas/DocumentResponse'
      '404':
        description: Document not found
    put:
      tags:
      - documents
      summary: Update a document
      description: Update an existing document
      operationId: updateDocument
      parameters:
      - name: id
        in: path
        description: Document ID
        required: true
        schema:
        type: integer
        format: int64
      requestBody:
      description: Updated document object
      content:
        application/json:
        schema:
          $ref: '#/components/schemas/Document'
      responses:
      '200':
        description: Document updated successfully
        content:
        application/json:
          schema:
          $ref: '#/components/schemas/Document'
      '404':
        description: Document not found
    delete:
      tags:
      - documents
      summary: Delete a document
      description: Delete a document by its ID
      operationId: deleteDocument
      parameters:
      - name: id
        in: path
        description: Document ID
        required: true
        schema:
        type: integer
        format: int64
      responses:
      '204':
        description: Document deleted successfully
      '404':
        description: Document not found
    /documents/group/{groupId}:
    get:
      tags:
      - documents
      summary: Get documents by group
      description: Retrieve all documents belonging to a specific group
      operationId: getAllDocumentsByGroup
      parameters:
      - name: groupId
        in: path
        description: Document group ID
        required: true
        schema:
        type: integer
        format: int64
      responses:
      '200':
        description: Group of documents retrieved successfully
        content:
        application/json:
          schema:
          type: array
          items:
            $ref: '#/components/schemas/DocumentMetadataResponse'
    /documents/user/{userId}:
    get:
      tags:
      - documents
      summary: Get documents by user
      description: Retrieve all documents for a specific user
      operationId: getDocumentsByUser
      parameters:
      - name: userId
        in: path
        description: User ID
        required: true
        schema:
        type: integer
        format: int64
      responses:
      '200':
        description: Documents retrieved successfully
        content:
        application/json:
          schema:
          type: array
          items:
            $ref: '#/components/schemas/Document'
    /documents/{id}/line-comments:
    get:
      tags:
      - documents
      summary: Get line comment counts
      description: Get comment counts per line for a document
      operationId: getLineCommentCounts
      parameters:
      - name: id
        in: path
        description: Document ID
        required: true
        schema:
        type: integer
        format: int64
      responses:
      '200':
        description: Line comment counts retrieved successfully
        content:
        application/json:
          schema:
          type: object
          properties:
            documentId:
            type: integer
            format: int64
            lineCommentCounts:
            type: object
            additionalProperties:
              type: integer
      '404':
        description: Document not found
...